+++ src/PhpSpreadsheet/Calculation/Database.php
@@ -34,7 +34,7 @@ class Database
      *                                        the column label in which you specify a condition for the
      *                                        column.
      *
-     * @return float|string
+     * @return null|float|string
      */
     public static function DAVERAGE($database, $field, $criteria)
     {