+++ src/PhpSpreadsheet/Calculation/Internal/WildcardMatch.php
@@ -5,9 +5,9 @@ namespace PhpOffice\PhpSpreadsheet\Calculation\Internal;
 class WildcardMatch
 {
     private const SEARCH_SET = [
-        '/([^~])(\*)/ui',
+        '/(?<!~)\*/ui',
         '/~\*/ui',
-        '/([^~])(\?)/ui',
+        '/(?<!~)\?/ui',
         '/~\?/ui',
     ];
 
@@ -20,6 +20,11 @@ class WildcardMatch
 
     public static function wildcard(string $wildcard): string
     {
+        // Preg Escape the wildcard, but protecting the Excel * and ? search characters
+        $wildcard = str_replace(['*', '?'], [0x1A, 0x1B], $wildcard);
+        $wildcard = preg_quote($wildcard);
+        $wildcard = str_replace([0x1A, 0x1B], ['*', '?'], $wildcard);
+
         return preg_replace(self::SEARCH_SET, self::REPLACEMENT_SET, $wildcard);
     }
 
@@ -29,6 +34,6 @@ class WildcardMatch
             return true;
         }
 
-        return (bool) preg_match("/{$wildcard}/ui", $value);
+        return (bool) preg_match("/^{$wildcard}\$/mui", $value);
     }
 }